gateway/requirements.txt
aniso8601==7.0.0
astroid==2.3.3
click==7.1.2
colorama==0.4.3
Flask==1.1.2
Flask-GraphQL==2.0.1
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
grpcio==1.30.0
grpcio-tools==1.30.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.8.0
mccabe==0.6.1
pathtools==0.1.2
promise==2.3
protobuf==3.12.2
pylint==2.4.4
python-dotenv==0.13.0
Rx==1.6.1
six==1.15.0
Werkzeug==1.0.1
wrapt==1.11.2
